+++ jest.config.ts
@@ -20,6 +20,7 @@ export default {
     '^@errors/(.*)$': '<rootDir>/src/errors/$1',
     '^@recoil/(.*)$': '<rootDir>/src/recoil/$1',
     '^@validation/(.*)$': '<rootDir>/src/validation/$1',
+    '^@queries/(.*)$': '<rootDir>/src/queries/$1',
   },
   testPathIgnorePatterns: ['<rootDir>/__tests__/constants/', '<rootDir>/__tests__/utils/'],
 };
