+++ src/queries/queryKey.ts
@@ -1,0 +1,9 @@
+import { ProductDropdownOptions } from '@components/product/ProductDropdown/ProductDropdown.type';
+
+export const QUERY_KEY = {
+  products: ['products'],
+  productsWithPagination: (dropdownOptions: ProductDropdownOptions) => [
+    ...QUERY_KEY.products,
+    dropdownOptions,
+  ],
+} as const;
